zakat360/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Script==2.0.6
SQLAlchemy==2.0.23
Werkzeug==2.3.7
Jinja2==3.1.2
APScheduler==3.10.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
reportlab==4.0.7
pytest==7.4.3
pytest-flask==1.3.0
Authlib==1.3.0
Flask-Babel==4.0.0